feat: network controller sends interrupts
To find out which interrupt is sent from PCI we have have
to parse pcmp which is already deprecated. The alternative
would be to parse and run some ACPI bytecode, which is too
complex.

One key learning was that one has to use the right outb/w/l
and inb/w/l function otherwise the device ignores it. That's
different from memory access, but it makes sense if I imagine
how a device gets the actual data, e.g. it has a pin for each
bit and only a few get current. Not sure if that even makes sense
but anyhow :)