cleaned up ps/2 related code
There were several notes in the code that there was a hanging bug. Turns
out this was because after enabling the ps/2 interrupts I tried to read
the config byte which resulted in an interrupt and from then on
something was off.

This is fixed. Additionally I learned that the ps/2 controller sends an
interrupt for each of the 3 bytes the mouse sends (1 mouse packet is 3
bytes). I tried to read them at once which is wrong or at least can be
wrong and result in hangs again.

Just relying on interrupts the wait functions are not needed anymore.

When configuring ps/2 make sure to disable the interrupts, etc.