move mouse and keyboard driver into tasks
We do some basic message passing to communicate data from the interrupt context
to the driver tasks.

There was a bug that the system became unresponsive when handling input. Turns
out printf in the schedule function took so much time that the schedule function
starved all other tasks for CPU. For now I removed the printf calls and also
increased the timer from 1ms to 10ms. Long term I need to reconsider when to
fire and reset the timer.