first user mode task doing 2 syscalls
This does
- paging in the user code and stack
- set up the correct GDT
- configure the CPU to allow syscall/sysret

It's really hacky and disables most of the old code to make it work.
It embeds an object file built with Bazel in tmp/ inside the kernel
binary using the linker script.

The paging code needs work. Physical memory needs to be page aligned but
the way we pass it in we don't know the length that will actually become
allocated befor calling pages_map_contiguous.