fix interrupts not working in user mode
If an interrupt happens while the CPU is in level 3 a stack
switch happens. This works by configuring the TSS and TSS descriptor.

The order in switch_task is important because once page tables are
switched a stack may become unusable. The iretq function will then cause
a page fault.